parse environment variable into an int

+++ landoapi/models/landing_job.py
@@ -14,7 +14,7 @@ from landoapi.storage import db
 
 logger = logging.getLogger(__name__)
 
-DEFAULT_GRACE_SECONDS = os.environ.get("DEFAULT_GRACE_SECONDS", 60 * 2)
+DEFAULT_GRACE_SECONDS = int(os.environ.get("DEFAULT_GRACE_SECONDS", 60 * 2))
 
 
 @enum.unique
